test: 인수 테스트 내부의 설정 관련 항목을 상위 클래스로 추출
- 실제 테스트 코드를 작성하는 하위 테스트 클래스는 테스트만을 담당하는 방식으로 리팩토링
